125 Years banner image
These files have the changes to put the 125 Years torch image in the
header of each page.  The current link takes you to the staging so
after it's all live that will need to be changed.  The image is right
aligned and so it can be replaced by any image and it will right align
in the same spot.