Navigation Bug Fix
When the accordian navigation would open the corresponding section for
the page suc as on this page:
http://www.byui.edu/agriculture-life-sciences/contact
It would slide down the nested ul element but it wouldn't add the
"open" class to the arrow which is what animates the arrow to rotate
showing that it is open.  I added a line to the accordianNav code to
add that class when the page loads.