New Feedback button image
This is the final feedback image that they wanted.  Both this image and
the other in this folder will eventually need to be hosted somewhere
and the html img tag's src attribute will have to be changed to point
to the images so that the html can be dropped on any server and still
work.