Fixes the contact info box as well as the facebook wall width problems
Problem can be seen here: www.byui.edu/newsroom

For whatever reason the width's that were set in the main.css file were
not correct or non existant.  #contact_info needed a width of 220px.
The .scrollbar needs to be at 218px so that with a 1px border it is
220px.  The .ff .scrollbar-pane needs to be 208px to take into account
the new .scrollbar width.  The .fb_wall must be 186px to take into
account the scrollbar.