JQuery Migrate and Analytics
Added the custom JQuery Migrate script.  Included it in the header
right after JQuery is loaded.  Changed JQuery to 1.10.2.  Also moved
the order of the s_code.js and omniture_code.js script tags.  s_code.js
has to be first or else the tracking script won't be loaded when
omniture_code.js tries to send information to Site Catalyst.